adding desc and thc amounts might have to redo the tables

+++ backend/data/migrations/20200816131608_create_tables.js
@@ -13,7 +13,7 @@ exports.up = function(knex) {
         table.string('product_name');
         table.integer('price');
         table.string('product_type');
-        table.integer('thc_amount');
+        table.float('thc_amount');
         table.string('image');
         table.string('description');
         table.integer('company_id')
